refactor: 객체 지향 원칙에 따라 UI 구조 개선
단일 책임 원칙(SRP)을 적용하여 main.py에 집중되어 있던 UI 관련 코드를 별도의 클래스와 파일로 분리합니다.

- `ui/` 디렉토리 생성
- `ControlPanelWidget`: 좌측 컨트롤 패널 UI 로직 분리
- `DataViewWidget`: 우측 데이터 뷰 UI 로직 분리
- `MainWindow`: 분리된 위젯들을 조립하고 메인 윈도우를 관리하는 역할로 단순화
